feat: support config to allow untrusted certs
Add new configuration to allow untrusted certificates, useful for testing but dangerious for
production so it needs to be configured explicitly.
