+++ Client.UnitTests/ZeebeClientTest.cs
@@ -60,191 +60,6 @@ namespace Zeebe.Client
         }
 
         [Test]
-        public async Task ShouldUseTransportEncryption()
-        {
-            // given
-            GrpcEnvironment.SetLogger(new ConsoleLogger());
-
-            var keyCertificatePairs = new List<KeyCertificatePair>();
-            var serverCert = File.ReadAllText(ServerCertPath);
-            keyCertificatePairs.Add(new KeyCertificatePair(serverCert, File.ReadAllText(ServerKeyPath)));
-            var channelCredentials = new SslServerCredentials(keyCertificatePairs);
-
-            var server = new Server();
-            server.Ports.Add(new ServerPort("localhost", 26505, channelCredentials));
-
-            var testService = new GatewayTestService();
-            var serviceDefinition = Gateway.BindService(testService);
-            server.Services.Add(serviceDefinition);
-            server.Start();
-
-            // client
-            var zeebeClient = ZeebeClient.Builder()
-                    .UseGatewayAddress("localhost:26505")
-                    .UseTransportEncryption(ClientCertPath)
-                    .Build();
-
-            // when
-            var publishMessageResponse = await zeebeClient
-                .NewPublishMessageCommand()
-                .MessageName("messageName")
-                .CorrelationKey("p-1")
-                .Send();
-
-            // then
-            Assert.NotNull(publishMessageResponse);
-        }
-
-        [Test]
-        public async Task ShouldUseTransportEncryptionWithServerCert()
-        {
-            // given
-            GrpcEnvironment.SetLogger(new ConsoleLogger());
-
-            var keyCertificatePairs = new List<KeyCertificatePair>();
-            var serverCert = File.ReadAllText(ServerCertPath);
-            keyCertificatePairs.Add(new KeyCertificatePair(serverCert, File.ReadAllText(ServerKeyPath)));
-            var channelCredentials = new SslServerCredentials(keyCertificatePairs);
-
-            var server = new Server();
-            server.Ports.Add(new ServerPort("localhost", 26505, channelCredentials));
-
-            var testService = new GatewayTestService();
-            var serviceDefinition = Gateway.BindService(testService);
-            server.Services.Add(serviceDefinition);
-            server.Start();
-
-            // client
-            var zeebeClient = ZeebeClient.Builder()
-                .UseGatewayAddress("localhost:26505")
-                .UseTransportEncryption(ServerCertPath)
-                .Build();
-
-            // when
-            var publishMessageResponse = await zeebeClient
-                .NewPublishMessageCommand()
-                .MessageName("messageName")
-                .CorrelationKey("p-1")
-                .Send();
-
-            // then
-            Assert.NotNull(publishMessageResponse);
-        }
-
-        [Test]
-        public async Task ShouldFailOnWrongCert()
-        {
-            // given
-            GrpcEnvironment.SetLogger(new ConsoleLogger());
-
-            var keyCertificatePairs = new List<KeyCertificatePair>();
-            var serverCert = File.ReadAllText(ServerCertPath);
-            keyCertificatePairs.Add(new KeyCertificatePair(serverCert, File.ReadAllText(ServerKeyPath)));
-            var channelCredentials = new SslServerCredentials(keyCertificatePairs);
-
-            var server = new Server();
-            server.Ports.Add(new ServerPort("localhost", 26505, channelCredentials));
-
-            var testService = new GatewayTestService();
-            var serviceDefinition = Gateway.BindService(testService);
-            server.Services.Add(serviceDefinition);
-            server.Start();
-
-            var zeebeClient = ZeebeClient.Builder()
-                .UseGatewayAddress("localhost:26505")
-                .UseTransportEncryption(WrongCertPath)
-                .Build();
-
-            // when
-            try
-            {
-                await zeebeClient
-                    .NewPublishMessageCommand()
-                    .MessageName("messageName")
-                    .CorrelationKey("p-1")
-                    .Send();
-                Assert.Fail();
-            }
-            catch (RpcException rpcException)
-            {
-                // expected
-                Assert.AreEqual(rpcException.Status.StatusCode, StatusCode.Unavailable);
-            }
-        }
-
-        [Test]
-        public async Task ShouldUseAccessToken()
-        {
-            // given
-            GrpcEnvironment.SetLogger(new ConsoleLogger());
-
-            var keyCertificatePairs = new List<KeyCertificatePair>();
-            var serverCert = File.ReadAllText(ServerCertPath);
-            keyCertificatePairs.Add(new KeyCertificatePair(serverCert, File.ReadAllText(ServerKeyPath)));
-            var channelCredentials = new SslServerCredentials(keyCertificatePairs);
-
-            var server = new Server();
-            server.Ports.Add(new ServerPort("localhost", 26505, channelCredentials));
-
-            var testService = new GatewayTestService();
-            var serviceDefinition = Gateway.BindService(testService);
-            server.Services.Add(serviceDefinition);
-            server.Start();
-
-            // client
-            var zeebeClient = ZeebeClient.Builder()
-                .UseGatewayAddress("localhost:26505")
-                .UseTransportEncryption(ClientCertPath)
-                .UseAccessToken("token")
-                .Build();
-
-            // when
-            await zeebeClient.TopologyRequest().Send();
-            await zeebeClient.TopologyRequest().Send();
-            var topology = await zeebeClient.TopologyRequest().Send();
-
-            // then
-            Assert.NotNull(topology);
-        }
-
-        [Test]
-        public async Task ShouldUseAccessTokenSupplier()
-        {
-            // given
-            GrpcEnvironment.SetLogger(new ConsoleLogger());
-
-            var keyCertificatePairs = new List<KeyCertificatePair>();
-            var serverCert = File.ReadAllText(ServerCertPath);
-            keyCertificatePairs.Add(new KeyCertificatePair(serverCert, File.ReadAllText(ServerKeyPath)));
-            var channelCredentials = new SslServerCredentials(keyCertificatePairs);
-
-            var server = new Server();
-            server.Ports.Add(new ServerPort("localhost", 26505, channelCredentials));
-
-            var testService = new GatewayTestService();
-            var serviceDefinition = Gateway.BindService(testService);
-            server.Services.Add(serviceDefinition);
-            server.Start();
-
-            // client
-            var accessTokenSupplier = new SimpleAccessTokenSupplier();
-            var zeebeClient = ZeebeClient.Builder()
-                .UseGatewayAddress("localhost:26505")
-                .UseTransportEncryption(ClientCertPath)
-                .UseAccessTokenSupplier(accessTokenSupplier)
-                .Build();
-
-            // when
-            await zeebeClient.TopologyRequest().Send();
-            await zeebeClient.TopologyRequest().Send();
-            var topology = await zeebeClient.TopologyRequest().Send();
-
-            // then
-            Assert.NotNull(topology);
-            Assert.AreEqual(3, accessTokenSupplier.Count);
-        }
-
-        [Test]
         public void ShouldCalculateNextGreaterWaitTime()
         {
             // given
@@ -269,19 +84,6 @@ namespace Zeebe.Client
 
             // then
             Assert.AreEqual(TimeSpan.FromSeconds(ZeebeClient.MaxWaitTimeInSeconds), maxTime);
-        }
-
-        private class SimpleAccessTokenSupplier : IAccessTokenSupplier
-        {
-            public int Count { get; private set; }
-
-            public Task<string> GetAccessTokenForRequestAsync(
-                string authUri = null,
-                CancellationToken cancellationToken = default(CancellationToken))
-            {
-                Count++;
-                return Task.FromResult("token");
-            }
         }
     }
 }
